+++ CHANGELOG.md
@@ -18,6 +18,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 * Fixed: compatibility with multiple the_content calls (Yoast)
 * Fixed: reset to showcontinued flag in Schedule (table view)
 * Added: show avatar and featured image URLs to Data API output
+* Added: option to ping Netmix directory on show updates
 
 ### 2.3.0
 * Include: Plugin Loader (1.1.0) with plugin options and settings